Improved performance when creating tons of SpanNear queries. Fixes issue #345.
The main performance difference in the test code was actually due to
a fix: in Whoosh 2.5, the Phrase query was fixed to call to_bytes() on
the phrase words. When making phrases 100s or 1000s of words long, the
difference was very noticeable.

Made two fixes to improve the performance:

1. Fixed logic in Phrase.matcher() to fail early on a missing word,
instead of calling to_bytes() on all words and THEN checking if
they're missing. This made a huge difference in the test code because
it was mostly searching for non-existant words.

2. Added a new SpanNear2 query type and changed Phrase to use it
instead of SpanNear. SpanNear2 has slightly less overhead.

These two changes make the test code actually run faster than in 2.4.